sheave_core::handlers
Trait MeasureAcknowledgement

pub trait MeasureAcknowledgement {
    // Required methods
    fn begin_measuring(&mut self);
    fn finish_measuring(&mut self);
    fn add_amount(&mut self, amount: u32);
    fn get_current_amount(&mut self) -> u32;

    // Provided method
    fn as_acknowledgement(&mut self) -> Acknowledgement { ... }
}

Measurement receiving chunk size.

Clients and the server are each required to send Acknowledgement messages when received chunk sizes exceed their bandwidths. This trait defines ways to measure receiving chunk sizes for any stream.

Required Methods

fn begin_measuring(&mut self)

Resets measured count and turns the flag into on, for beginning measurement.

fn finish_measuring(&mut self)

Resets measured count and turns the flag into off, for finishing measurement.

fn add_amount(&mut self, amount: u32)

Adds received chunk size to current one.

fn get_current_amount(&mut self) -> u32

Gets current chunk size.

Provided Methods

fn as_acknowledgement(&mut self) -> Acknowledgement

Wraps current chunk size into an Acknowledgement message.