Module sheave_core::messages::headers

The Chunk Headers

Every chunk has following headers:

The Basic Header

The Message Header

Extended Timestamp

The Basic Header

This indicates:

A pattern of followed message header.

Which the chunk stream are we in.

The Message Header

This indicates:

A timestamp which has passed from its stream started.

A pattern of followed chunk data.

Which the message stream are we in.

This must be depended its format on a value in the basic header. That is, the message header has following correspondence with the basic header:

|Value in the basic header|Format of the message header| | -: | -: | |0|11 bytes| |1|7 bytes| |2|3 bytes| |3|0 bytes|

Extended Timestamp

This is added when a timestamp in its message header exceeded the 3 bytes range. In that case, note its field must be filled with 0xFFFFFF (16777215).

Re-exports

pub use self::message::New;

pub use self::message::SameSource;

pub use self::message::TimerChange;

Structs

BasicHeader

Indicates the chunk stream and message header’s format. This header has 3 types.

Enums

MessageFormat

The first 2 bits to indicate a format of message header.

MessageHeader

Indicates a chunk datum format and which stream is it into. This header has 4 types.

MessageType

Representation of message type id byte of the Message header.