sheave_core::messages::headers

Struct MessageHeader

pub struct MessageHeader { /* private fields */ }

Indicates a chunk datum format and which stream is it into.

This header has 4 types.

Total	Timestamp	Message Length	Message Type	Message ID

11	3	3	1	4

7	3	3	1

3	3

0

The unit of every length item is bytes.

11 bytes type is required when a new message ID is necessary, that is, communicating with a partner on a new message stream.

7 bytes type is required when a message which is different either type or length, sends/receives on same message stream.

3 bytes type is required when a message which is same both type and length, sends/receives on same message stream.

0 bytes type is required when concatenates a message which is same both type and length, but exceeding the chunk size on same message stream.

Note that 0 bytes type is required to consider of the message length because can be contained into its chunk datum.

tuples are convertible into message headers.

Tuples	Message Header Format

(Duration, u32, u8, u32)	New

`(Duration, u32, u8)	SameSource

Duration	TimerChange

()	Continue

Examples

use std::time::Duration;
use sheave_core::messages::headers::MessageHeader;

let new: MessageHeader = (Duration::default(), u32::default(), u8::default().into(), u32::default()).into();
assert!(new.get_message_id().is_some());

let same_source: MessageHeader = (Duration::default(), u32::default(), u8::default().into()).into();
assert!(same_source.get_message_length().is_some());
assert!(same_source.get_message_type().is_some());

let timer_change: MessageHeader = Duration::default().into();
assert!(timer_change.get_timestamp().is_some());

let cont: MessageHeader = ().into();
assert!(cont.get_timestamp().is_none())

Implementations

impl MessageHeader

pub fn get_timestamp(&self) -> Option<Duration>

Gets a timestamp.

Only 0 bytes type returns None.

Examples

use std::time::Duration;
use sheave_core::messages::headers::MessageHeader;

// In case of 11 bytes type.
let new: MessageHeader = (Duration::default(), u32::default(), u8::default().into(), u32::default()).into();
assert!(new.get_timestamp().is_some());

// In case of 7 bytes type.
let same_source: MessageHeader = (Duration::default(), u32::default(), u8::default().into()).into();
assert!(same_source.get_timestamp().is_some());

// In case of 3 bytes type.
let timer_change: MessageHeader = Duration::default().into();
assert!(timer_change.get_timestamp().is_some());
 
// In case of 0 bytes type.
let cont: MessageHeader = ().into();
assert!(cont.get_timestamp().is_none())

pub fn get_message_length(&self) -> Option<u32>

Gets a message length.

0 bytes type and 3 bytes type return None.

Examples

use std::time::Duration;
use sheave_core::messages::headers::MessageHeader;

// In case of 11 bytes type.
let new: MessageHeader = (Duration::default(), u32::default(), u8::default().into(), u32::default()).into();
assert!(new.get_message_length().is_some());

// In case of 7 bytes type.
let same_source: MessageHeader = (Duration::default(), u32::default(), u8::default().into()).into();
assert!(same_source.get_message_length().is_some());

// In case of 3 bytes type.
let timer_change: MessageHeader = Duration::default().into();
assert!(timer_change.get_message_length().is_none());

// In case of 0 bytes type.
let cont: MessageHeader = ().into();
assert!(cont.get_message_length().is_none())

pub fn get_message_type(&self) -> Option<MessageType>

Gets a message type.

0 bytes type and 3 bytes type return None.

Examples

use std::time::Duration;
use sheave_core::messages::headers::MessageHeader;

// In case of 11 bytes type.
let new: MessageHeader = (Duration::default(), u32::default(), u8::default().into(), u32::default()).into();
assert!(new.get_message_type().is_some());

// In case of 7 bytes type.
let same_source: MessageHeader = (Duration::default(), u32::default(), u8::default().into()).into();
assert!(same_source.get_message_type().is_some());

// In case of 3 bytes type.
let timer_change: MessageHeader = Duration::default().into();
assert!(timer_change.get_message_type().is_none());

// In case of 0 bytes type.
let cont: MessageHeader = ().into();
assert!(cont.get_message_type().is_none())

pub fn get_message_id(&self) -> Option<u32>

Gets a message ID.

All but 11 bytes type returns None.

Examples

use std::time::Duration;
use sheave_core::messages::headers::MessageHeader;

// In case of 11 bytes type.
let new: MessageHeader = (Duration::default(), u32::default(), u8::default().into(), u32::default()).into();
assert!(new.get_message_id().is_some());

// In case of 7 bytes type.
let same_source: MessageHeader = (Duration::default(), u32::default(), u8::default().into()).into();
assert!(same_source.get_message_id().is_none());

// In case of 3 bytes type.
let timer_change: MessageHeader = Duration::default().into();
assert!(timer_change.get_message_id().is_none());

// In case of 0 bytes type.
let cont: MessageHeader = ().into();
assert!(cont.get_message_id().is_none())