Struct sheave_core::messages::PeerBandwidth

pub struct PeerBandwidth(/* private fields */);

The message to tell the client-side bandwidth. This has 2 ways of comparision by which field you specifies.

Examples

use sheave_core::messages::{
    LimitType,
    PeerBandwidth,
};

let peer_bandwidth = PeerBandwidth::default();

// When you compare this message with a bandwidth number.
assert!(2500000u32 == peer_bandwidth);
assert!(0 <= peer_bandwidth);
// When you compare this message with a limit type.
assert!(LimitType::default() == peer_bandwidth)

Implementations

impl PeerBandwidth

pub fn new(peer_bandwidth: u32, limit_type: LimitType) -> Self

Constructs a PeerBandwidth message.

pub fn get_inner_bandwidth(&self) -> u32

pub fn get_inner_limit_type(&self) -> LimitType