sheave_core::messages

Struct Play

pub struct Play { /* private fields */ }

The command to tell playing information.

Following format is required:

Field	AMF Type	Value

Null	Nothing but an AMF’s type marker is in.

Stream Name	String	A name for subscribing a data from the server.

Start Time	Number	Time offset of data subscribing.
Note this can be several negative number (See Play Mode).

Play Mode

Value	Play Mode	Description

-2	Both	Subscribes recorded data if its data isn’t on a livestream (default).

-1	Live	Subscribes only as livestream.

0
(And above)	Recorded	Subscribes only as recorded data.

Note the server can treat as some error if its data doesn’t exist as specified mode.

Implementations

impl Play

pub fn new(stream_name: AmfString, start_time: Number) -> Self

Constructs a Play command.

pub fn get_stream_name(&self) -> &AmfString

Gets the stream name. (e.g. filename)

pub fn get_start_time(&self) -> Number

Gets the start time.

pub fn get_play_mode(&self) -> PlayMode

Gets the play mode.