sheave_core::messages

Struct Publish

pub struct Publish { /* private fields */ }

The command to tell publishing information.

Following format is required:

Field	AMF Type	Value

Null	Nothing but an AMFโs type marker is in.

Publishing Name	String	A name for publishing a data to the server.

Publishing Type	String	See Publishing Type.

Publishing Type

The publish command requires you to specify one of โPublishing Typeโ in its request. Publishing Type means:

Pubishing Type	Description

"live"	Only streaming.
Its data will never be stored.

"record"	Its data will be stored.
If publishing name duplicated, it is rewritten as a new file.

"append"	Same as "record" excepts is appended its data if publishing name duplicated.

Implementations

impl Publish

pub fn new(publishing_name: AmfString, publishing_type: AmfString) -> Self

Constructs a Publish command.

pub fn get_publishing_name(&self) -> &AmfString

Gets the publishing identifier. (e.g. filename)

pub fn get_publishing_type(&self) -> &AmfString

Gets one of publishing type which is either "live", "record" or "append".