sheave_core::messages

Struct WindowAcknowledgementSize

pub struct WindowAcknowledgementSize(/* private fields */);

The message to tell the bandwidth limit as a receiver side.

Following format is required.

Field	Length (in bytes)	Description

Bandwidth	4	The bandwidth number of a receiver side.

When some amount of receiving data got exceeded this bandwidth, a receiver must send about it to a sender, by the Acknowledgement message.

Almost RTMP tools are specifying it 2500000 in bits as their defaults.

Implementations

impl WindowAcknowledgementSize

pub fn new(window_acknowledgement_size: u32) -> Self

Constructs an WindowAcknowledgementSize message.

pub fn get_inner(&self) -> u32

Gets an inner value.