Trait sheave_core::messages::ChunkData

pub trait ChunkData {
    const CHANNEL: Channel;
    const MESSAGE_TYPE: MessageType;
}

Uniforms chunk data by a channel numbers and a message type.

When writing into streams, we are required to imprint a chunk ID and a message type ID into their headers. This makes you to reduce its cost. For example, write_chunk use this for writing chunks correctly.

Required Associated Constants

const CHANNEL: Channel

const MESSAGE_TYPE: MessageType

Object Safety

This trait is not object safe.