Trait sheave_core::messages::Command

pub trait Command {
    // Required methods
    fn get_command_name(&self) -> &str;
    fn get_transaction_id(&self) -> Number;
}

The accessor for common fields in every command chunk.

All command chunk but onFCPublish has its command name and identifier for the transaction (only onFCPublish has just its command name). This provide ways for accessing to their common fields of command chunks uniformly. For example, decode and encode which are implemented for them use this for reading them from streams or writing them into streams.

Required Methods

fn get_command_name(&self) -> &str

fn get_transaction_id(&self) -> Number