sheave_core::messages

Trait UserControl

pub trait UserControl {
    const EVENT_TYPE: EventType;
}

Uniforms user control messages by an event type.

When reading from streams or writing into streams, we are required to imprint a event type ID into their messages. This makes you to reduce its cost.

Required Associated Constants

const EVENT_TYPE: EventType

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.