Struct sheave_core::net::rtmp::RtmpStream

pub struct RtmpStream { /* private fields */ }

A stream for RTMP that wrapped Tokio’s TcpStream. If you constructs this struct from some address, use RtmpStream::connect("aaa.bbb.ccc,ddd:1935"). If you do it from already created std’s TCPStream. use RtmpStream::from_std(std_stream)

Implementations

impl RtmpStream

pub async fn connect<A: ToSocketAddrs>(addr: A) -> IOResult<Self>

Opens a RTMP connection to a remote host. When connection succeeded, this wraps tokio’s TcpStream in RtmpStream. Read more

pub fn from_std(std_stream: StdStream) -> IOResult<Self>

Creates new RtmpStream from a std::net::TcpStream. When connection succeeded, this wraps tokio’s TcpStream in RtmpStream. Read more

pub fn into_std(self) -> IOResult<StdStream>

Turns a sheave_core::net::rtmp::RtmpStream into std::net::TcpStream`. Read more

pub fn local_addr(&self) -> IOResult<SocketAddr>

Returns the local address that this stream is bound to. Read more

pub fn take_error(&self) -> IOResult<Option<IOError>>

Returns the value of the SO_ERROR option.

pub fn peer_addr(&self) -> IOResult<SocketAddr>

Returns the remote address that this stream is connected to. Read more

pub fn poll_peek( &self, cx: &mut FutureContext<'_>, buf: &mut ReadBuf<'_> ) -> Poll<IOResult<usize>>

Attempts to receive data on the socket, without removing that data from the queue, registering the current task for wakeup if data is not yet available. Read more

pub async fn ready(&self, interest: Interest) -> IOResult<Ready>

Waits for any of the requested ready states. Read more

pub async fn readable(&self) -> IOResult<()>

Waits for the socket to become readable. Read more

pub fn poll_read_ready(&self, cx: &mut FutureContext<'_>) -> Poll<IOResult<()>>

Polls for read readiness. Read more

pub fn try_read(&self, buf: &mut [u8]) -> IOResult<usize>

Tries to read data from the stream into the provided buffer, returning how many bytes were read. Read more

pub fn try_read_vectored(&self, bufs: &mut [IoSliceMut<'_>]) -> IOResult<usize>

Tries to read data from the stream into the provided buffers, returning how many bytes were read. Read more

pub fn try_read_buf<B: BufMut>(&self, buf: &mut B) -> IOResult<usize>

Tries to read data from the stream into the provided buffer, advancing the buffer’s internal cursor, returning how many bytes were read. Read more

pub async fn writable(&self) -> IOResult<()>

Waits for the socket to become writable. Read more

pub fn poll_write_ready(&self, cx: &mut FutureContext<'_>) -> Poll<IOResult<()>>

Polls for write readiness. Read more

pub fn try_write(&self, buf: &[u8]) -> IOResult<usize>

Tries to write several buffers to the stream, returning how many bytes were written. Read more

pub fn try_write_vectored(&self, bufs: &[IoSlice<'_>]) -> IOResult<usize>

Tries to write several buffers to the stream, returning how many bytes were written. Read more

pub fn try_io<R>( &self, interest: Interest, f: impl FnOnce() -> IOResult<R> ) -> IOResult<R>

Tries to read or write from the socket using a user-provided IO operation. Read more

pub async fn async_io<R>( &self, interest: Interest, f: impl FnMut() -> IOResult<R> ) -> IOResult<R>

Reads or writes from the socket using a user-provided IO operation. Read more

pub async fn peek(&self, buf: &mut [u8]) -> IOResult<usize>

Receives data on the socket from the remote address to which it is connected, without removing that data from the queue. On success, returns the number of bytes peeked. Read more

pub fn nodelay(&self) -> IOResult<bool>

Gets the value of the TCP_NODELAY option on this socket. Read more

pub fn set_nodelay(&self, nodelay: bool) -> IOResult<()>

Sets the value of the TCP_NODELAY option on this socket. Read more

pub fn linger(&self) -> IOResult<Option<Duration>>

Reads the linger duration for this socket by getting the SO_LINGER option. Read more

pub fn set_linger(&self, dur: Option<Duration>) -> IOResult<()>

Sets the linger duration of this socket by setting the SO_LINGER option. Read more

pub fn ttl(&self) -> IOResult<u32>

Gets the value of the IP_TTL option for this socket. Read more

pub fn set_ttl(&self, ttl: u32) -> IOResult<()>

Sets the value for the IP_TTL option on this socket. Read more

pub fn split<'a>(&'a mut self) -> (ReadHalf<'a>, WriteHalf<'a>)

Splits a TcpStream into a read half and a write half, which can be used to read and write the stream concurrently. Read more

pub fn into_split(self) -> (OwnedReadHalf, OwnedWriteHalf)

Splits a TcpStream into a read half and a write half, which can be used to read and write the stream concurrently. Read more