sheave_core
Module writers

Functions

write_basic_header

Writes a basic header into streams.

write_chunk

A wrapper for writing a chunk into streams. The RTMP needs to refer previous states for deciding sending chunk pattern. But to check them in every step is troublesome and also can make some bug. This reduces their risks.

write_chunk_data

Writes a chunk data into streams.

write_encryption_algorithm

Writes one byte to indicate the encryption algorithm into streams.

write_extended_timestamp

Writes an extended timestramp into streams.

write_handshake

Writes actual handshake data into streams.

write_message_header

Writes a message header into streams.