Crate sheave_core

Modules

cli

The common CLI options.

handlers

Handling RTMP connections and data streaming.

handshake

Types for the handshake step in RTMP.

messages

The RTMP Messages

net

readers

writers

Macros

ecma_array

Constructs an ECMA array.

object

Constructs an AMF’s Object.

Structs

ByteBuffer

The stream buffer for encoding/decoding chunk data.

InsufficientBufferLength

An error that buffer has been empty during decoding chunks.

Traits

Decoder

Decodes bytes into a chunk data. This can return errors for decoding fails by something cause.

Encoder

Encodes a chunk data into bytes.

Functions

insufficient_buffer_length

A utility function of constructing an InsufficientBufferLength error.