sheave_core
Module handlers

Handling RTMP connections and data streaming.

Currently following handlers have been implemented.

Handshake

connect

releaseStream

FCPublish

createStream

publish

Structs

InconsistentSha

Tells that either digests or signatures are inconsistent in the handshake step.

LastChunk

The chunk information which is sent/received last.

RtmpContext

RTMP’s common contexts.

StreamGotExhausted

An error that stream has no data to write. Note this is currently used as to mean sucessful termination.

StreamWrapper

The wrapper for stream types.

VecStream

The simple stream for std’s buffer-like types.

Enums

PublisherStatus

Traits

AsyncHandler

The interface for handling RTMP connection steps with Future.

AsyncHandlerExt

The extension methods for handlers.

ErrorHandler

HandlerConstructor

The interface for providing the way to construct any handler to clients/servers.

MeasureAcknowledgement

Measurement receiving chunk size.

Middleware

Functions

inconsistent_sha

A utility function for wrapping the error InconsistentSha into ::std::io::Error.

stream_got_exhausted

A utility function of constructing an StreamGotExhausted