sheave_core

Module handlers

Structs

InconsistentSha

Tells that either digests or signatures is inconsistent in the handshake step.

LastChunk

The chunk information which is sent/received last.

RtmpContext

RTMP’s common contexts.

StreamGotExhausted

An error that stream has no data to write.

StreamWrapper

The wrapper for stream types.

VecStream

The simple stream for std’s buffer-like types.

Enums

ClientType

The pattern to distinguish whether client requests are publishers or subscribers.

PublisherStatus

The pattern of communication status between servers and publisher clients.

SubscriberStatus

The pattern of communication status between servers and subscriber clients.

Traits

AsyncHandler

The interface for handling RTMP connection steps with Future.

AsyncHandlerExt

The extension methods for handlers.

ErrorHandler

HandlerConstructor

The interface for providing the way to construct any handler to clients/servers.

MeasureAcknowledgement

Measurement receiving chunk size.

Middleware

Functions

inconsistent_sha

A utility function for constructing an InconsistentSha error.

stream_got_exhausted

A utility function of constructing an StreamGotExhausted error.