sheave_core::handlers

Struct RtmpContext

pub struct RtmpContext { /* private fields */ }

RTMP’s common contexts.

Many fields are optional by default. Because these data need for both of client and server but are sent/received later. Therefore the default constructor has been prepared instead of such as new.

Examples

use sheave_core::handlers::RtmpContext;

// When you create this struct.
RtmpContext::default();

Implementations

impl RtmpContext

pub fn make_weak_mut<'a>(self: &'a Arc<Self>) -> &'a mut Self

Gets a mutable reference via this wrapped by Arc.

Sheave uses this after wrapping into Arc. Because of making this shareable between every handling steps.

pub fn set_signed(&mut self, signed: bool)

Stores a flag to mean this handshake is signed.

pub fn is_signed(&mut self) -> bool

Indicates whether the handshake is signed.

pub fn set_receiving_chunk_size(&mut self, chunk_size: ChunkSize)

Sets a chunking size which reads from streams.

pub fn get_receiving_chunk_size(&mut self) -> ChunkSize

Gets a chunking size which reads from streams.

pub fn set_sending_chunk_size(&mut self, chunk_size: ChunkSize)

Sets a chunking size which writes into streams.

pub fn get_sending_chunk_size(&mut self) -> ChunkSize

Gets a chunkign size which writes into stream.

pub fn set_window_acknowledgement_size( &mut self, window_acknowledgement_size: WindowAcknowledgementSize, )

Sets the window acknowledgement size.

pub fn get_window_acknowledgement_size(&mut self) -> WindowAcknowledgementSize

Gets the window acknowledgement size.

pub fn set_peer_bandwidth(&mut self, peer_bandwidth: PeerBandwidth)

Sets the peer bandwidth.

pub fn get_peer_bandwidth(&mut self) -> PeerBandwidth

Gets the peer bandwidth.

pub fn set_buffer_length(&mut self, buffer_length: u32)

Sets the buffer length.

pub fn get_buffer_length(&mut self) -> u32

Gets the buffer length.

pub fn set_transaction_id(&mut self, transaction_id: Number)

Sets a transaction ID.

Mainly, this is used by server side contexts. Because of servers should echo same transaction ID in its response.

pub fn get_transaction_id(&mut self) -> Number

Gets a transaction ID sent.

pub fn increase_transaction_id(&mut self)

Increases current transaction ID.

Mainly, this is used by client side contexts. Because of clients should count which transaction is it now on.

pub fn set_database_url(&mut self, database_url: &str)

Sets the database url.

pub fn get_database_url(&mut self) -> Option<&String>

Gets the database url.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_database_url().is_none())

pub fn set_storage_path(&mut self, storage_path: &str)

Sets the storage path.

pub fn get_storage_path(&mut self) -> Option<&String>

Gets the storage path.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_storage_path().is_none())

pub fn set_client_addr(&mut self, client_addr: SocketAddr)

Sets a client IP address.

pub fn get_client_addr(&mut self) -> Option<SocketAddr>

Gets a client IP address.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_client_addr().is_none())

pub fn set_app(&mut self, app: &str)

Sets the app name.

This can be contained in a request URI of RTMP.

pub fn get_app(&mut self) -> Option<&AmfString>

Gets the app name.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_app().is_none())

pub fn set_topic_id(&mut self, topic_id: AmfString)

Sets a topic_id (e.g. filename) sent from a client.

pub fn reset_topic_id(&mut self)

Resets a topic_id from this context.

This is prepared for deleting a topic_id when receives the FCUnpublish command.

pub fn get_topic_id(&mut self) -> Option<&AmfString>

Gets a topic_id (e.g. filename) sent from a client.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_topic_id().is_none())

pub fn set_tc_url(&mut self, tc_url: &str)

Sets the tcUrl. This is a full URL in the RTMP request like following form.

rtmp://hostname/[app]/[topic_path]

Where app and topic_path can be unspecified. Clients specify above URL at the start of RTMP requests. Then the server checks app and topic_path in client-side Connect commands (if they are specified).

pub fn get_tc_url(&mut self) -> Option<&AmfString>

Gets the tcUrl.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_tc_url().is_none())

pub fn set_command_name(&mut self, command_name: AmfString)

Sets last command name.

pub fn get_command_name(&mut self) -> Option<&AmfString>

Gets last command name.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_command_name().is_none())

pub fn set_client_type(&mut self, client_type: ClientType)

Sets that its client is either publisher or subscriber.

Curently, the server distinguishes this by referring specific field in a command object which a connect command has. e.g. “fpad”, “capabilities” etc.

pub fn get_client_type(&mut self) -> Option<ClientType>

Gets the type that its client is which either publisher or subscriber.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_client_type().is_none())

pub fn set_publisher_status(&mut self, publisher_status: PublisherStatus)

Sets one of publisher’s status.

pub fn get_publisher_status(&mut self) -> Option<PublisherStatus>

Gets one of publisher’s status.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_publisher_status().is_none())

pub fn set_subscriber_status(&mut self, subscriber_status: SubscriberStatus)

Sets one of subscriber’s status.

pub fn get_subscriber_status(&mut self) -> Option<SubscriberStatus>

Gets one of subscriber’s status.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_subscriber_status().is_none())

pub fn set_encryption_algorithm( &mut self, encryption_algorithm: EncryptionAlgorithm, )

Stores the algorithm to encrypt this handshake.

pub fn get_encryption_algorithm(&mut self) -> Option<EncryptionAlgorithm>

Gets specieifed algorithm to encrypt this handshake.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_encryption_algorithm().is_none())

pub fn set_client_handshake(&mut self, handshake: Handshake)

Stores a cleint-side handshake bytes.

pub fn get_client_handshake(&mut self) -> Option<&Handshake>

Gets a client-side handshake bytes.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_client_handshake().is_none())

pub fn get_client_handshake_mut(&mut self) -> Option<&mut Handshake>

Gets a client-side handshake bytes as mutable.

Note:

This is currently used for only testing (also intagration tests contained).

This can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_client_handshake_mut().is_none())

pub fn set_server_handshake(&mut self, handshake: Handshake)

Stores a server-side handshake bytes.

pub fn get_server_handshake(&mut self) -> Option<&Handshake>

Gets a server-side handshake bytes.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_server_handshake().is_none())

pub fn set_command_object(&mut self, command_object: Object)

Sets a command object sent from a client.

pub fn get_command_object(&mut self) -> Option<&Object>

Gets a command object sent from a client.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_command_object().is_none())

pub fn set_properties(&mut self, properties: Object)

Sets a properties object of a server.

pub fn get_properties(&mut self) -> Option<&Object>

Gets a properties object of a server.

Note this can return None. e.g. When this field is dafault as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_properties().is_none())

pub fn set_information(&mut self, information: Object)

Sets a information object of a server.

pub fn get_information(&mut self) -> Option<&Object>

Gets a information object of a server.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_information().is_none())

pub fn set_message_id(&mut self, message_id: u32)

Sets a message ID of this stream.

pub fn reset_message_id(&mut self)

Resets a message ID from this context.

This is prepared for deleting tne message_id when receives the deleteStream command.

pub fn get_message_id(&mut self) -> Option<u32>

Gets a message ID of this stream.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_message_id().is_none())

pub fn set_playlist(&mut self, playlist: EcmaArray)

Sets a Playlist.

Currently, this is sent from several client like OBS.

pub fn get_playlist(&mut self) -> Option<&EcmaArray>

Gets a playlist.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_playlist().is_none())

Currently, this is sent from several client like OBS.

pub fn set_publishing_name(&mut self, publishing_name: AmfString)

Sets a publishing name of this stream.

pub fn get_publishing_name(&mut self) -> Option<&AmfString>

Gets a publishing name of this stream.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_publishing_name().is_none())

pub fn set_publishing_type(&mut self, publishing_type: AmfString)

Sets a publishing type of this stream.

pub fn get_publishing_type(&mut self) -> Option<&AmfString>

Gets a publishing type of this stream.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_publishing_type().is_none())

pub fn set_stream_name(&mut self, stream_name: AmfString)

Sets a stream name of this stream.

pub fn get_stream_name(&mut self) -> Option<&AmfString>

Gets a stream name of this stream.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_stream_name().is_none())

pub fn set_start_time(&mut self, start_time: Option<Duration>)

Sets a start time of this stream.

pub fn get_start_time(&mut self) -> Option<Duration>

Gets a start time of this stream.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_start_time().is_none())

pub fn set_play_mode(&mut self, play_mode: PlayMode)

Sets a play mode of this stream.

pub fn get_play_mode(&mut self) -> Option<PlayMode>

Gets a play mode of this stream.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_play_mode().is_none())

pub fn set_await_duration(&mut self, await_duration: Duration)

Sets a duration for awaiting of receiving some message.

Currently, this is used only clients during publishing audio/video data.

pub fn get_await_duration(&mut self) -> Option<Duration>

Gets a duration for awaiting of receiving some message.

Currently, this is used only clients during publishing audio/video data.

pub fn set_topic(&mut self, topic: Flv)

Sets a topic file/device.

pub fn get_topic(&mut self) -> Option<&Flv>

Gets a topic file/device.

Note this can return None. e.g. When this field is default as it is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_topic().is_none())

pub fn get_topic_mut(&mut self) -> Option<&mut Flv>

Gets a topic file/device as mutable.

Note this can return None. e.g. When it is the default as is.

Examples

use sheave_core::handlers::RtmpContext;

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_topic_mut().is_none())

pub fn insert_received_chunk(&mut self, chunk_id: u16, last_chunk: LastChunk)

Stores a last received chunk.

pub fn get_last_received_chunk(&mut self, chunk_id: &u16) -> Option<&LastChunk>

Loads a last received chunk.

If no last chunk is stored associated with specified ID, this returns None.

Examples

use sheave_core::{
    handlers::RtmpContext,
    messages::Channel
};

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_last_received_chunk(&Channel::System.into()).is_none())

pub fn get_last_received_chunk_mut( &mut self, chunk_id: &u16, ) -> Option<&mut LastChunk>

Loads a last received chunk as mutable.

If no last chunk is stored associated with specified ID, this returns None.

Examples

use sheave_core::{
    handlers::RtmpContext,
    messages::Channel
};

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_last_received_chunk_mut(&Channel::System.into()).is_none())

pub fn insert_sent_chunk(&mut self, chunk_id: u16, last_chunk: LastChunk)

Stores a last sent chunk.

pub fn get_last_sent_chunk(&mut self, chunk_id: &u16) -> Option<&LastChunk>

Loads a last sent chunk.

If no last chunk is stored associated with specified ID, this returns None.

Examples

use sheave_core::{
    handlers::RtmpContext,
    messages::Channel
};

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_last_sent_chunk(&Channel::System.into()).is_none())

pub fn get_last_sent_chunk_mut( &mut self, chunk_id: &u16, ) -> Option<&mut LastChunk>

Loads a last sent chunk as mutable.

If no last chunk is stored associated with specified ID, this returns None.

Examples

use sheave_core::{
    handlers::RtmpContext,
    messages::Channel
};

let mut rtmp_context = RtmpContext::default();
assert!(rtmp_context.get_last_sent_chunk_mut(&Channel::System.into()).is_none())