Module sheave_core::messages

The RTMP Messages

After handshake, both are required to negosiate what are needed for sending/receiving. It consists of following steps:

Connect

Connect

Exchanges informations about their applications each other. In this case, next format is required:

Field	AMF Type	Value

Command Name	String	"connect".

Transaction ID	Number	1

Command Object	Object	See Command Object.

In the command message, we negotiate by using the Action Message Format (v0).

Command Object

The Command Object is written detailed informations of both applications. Concretely, several of following pair is exchanged:

Key	AMF Type	Description

app	String	Server application name.

flashVer	String	Either a flash player version(client) or an FLV encoder version(server).

swfUrl	String	The source SWF file URL.

tcURL	String	The URL to connect server as the TCP. app value is included in this value.

fpad	Boolean	Whether a proxy is used.

audioCodecs	Number	Supported audio codecs.

videoCodecs	Number	Supported video codecs.

videoFunctions	Number	Supported video functions.

pageUrl	String	The URL of web page which SWF file is loaded.

objectEncoding	Number	A version of the Action Message Format.

Note: Something not in above can be exchanged.

Every chunk has the chunk headers. See sheave_core::messages::headers about them.

Modules

amf

The Action Message Formats

headers

The Chunk Headers

Structs

Connect

The command to tell the information that where connects from/to.

InconsistentCommand

An error means that some command name differs you expect.

Functions

inconsistent_command

A utility function of constructing an InconsistentCommand error.