sheave_core::messages::amf

Module v0

The AMF Data Types (version 0).

These are data types which are defined in the Action Message Format version 0 specification. Currently the RTMP uses following types:

Marker	AMF Data Type	Description

0	Number	The IEEE 754 double precision floating point number.

1	Boolean	The boolean that is represented as a 1 byte data. (C-like)

2	AmfString	The string that is limited its length the range of 2 bytes.

3	Object	The key/value-paired object that its value type is flexible.

5	Null	Only the marker. Any value doesn’t contain.

8	EcmaArray	Same as the Object except this has its length.

9	Object End	Indicates a stream of object terminaites there.

These are checked whether matched its marker with an actual data, by the receiver-side. Therefore any stream which contains any unmatched marker will be considered it’s invalid. Note that a marker of the Object End must appear only in AMF’s Object. Following descriptions are the actual formats of AMF.

Number

Marker (1 byte)

An IEEE 754 double precision value (8 bytes)

Boolean

Marker (1 byte)

A boolean value (1 byte)

String

Marker (1 byte)

Length (2 bytes)

UTF-8 characters (variable)

Object

Marker (1 byte)

Arbitrary number key-value pairs (variable)

The Object End marker that is associated with an empty string.

Note keys are without markers of the AMF String. To be exact, these aren’t defined as the AMF String in the specification, that is, these are just strings with lengths (within 2 bytes) at the beginning.

Next, values are limited just in the AMF data types. Currently, following AMF data types to be used in the RTMP are allowed to insert into the AMF Object type:

Number

Boolean

AmfString

Object

Null

EcmaArray

Null

Marker (1 byte)

EcmaArray

Marker (1 byte)

Length (4 bytes)

Arbitrary number key-value pairs (variable)

The Object End marker that is associated with an empty string.

Structs

AmfString

The UTF-8 string of AMF data types.

Boolean

The boolean representation of AMF data types. This uses 1 byte unsigned integer as a boolean value. Usually, 0 is treated as false, else is true.

EcmaArray

The sized object type of AMF.

Null

The value to mean that has no content.

Number

The IEEE 754 double precision floating point number of AMF data types.

Object

The anonymous object type of AMF. This consists of pairs of string keys and any AMF data types.

Enums

Marker

Representation of markers of the AMF data types.